pyproject.toml
[project]
name = "sgames"
version = "0.1.0"
description = "Information Retrieval: Search Engine for a scraped videogames collection"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckdb>=1.4.2",
    "flask>=3.1.2",
    "httpx>=0.28.1",
    "nltk>=3.9.2",
    "parsel>=1.10.0",
    "playwright>=1.56.0",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "scrapy>=2.13.3",
]

[project.scripts]
scraper = "scraper.main:main"
app = "app.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scraper", "app", "utils"]
